requirements.txt
opencv-python
fer
pyserial
tensorflow